chore: move grouped translate tests into their own file
