+++ internal/dataplane/parser/translate_httproute.go
@@ -35,10 +35,6 @@ func (p *Parser) ingressRulesFromHTTPRoutes() ingressRules {
 
 	if len(errs) > 0 {
 		for _, err := range errs {
-			// we log the error here instead of returning it for historical reasons
-			// and because this allows other HTTPRoute objects to be resolved. This
-			// loop structure was common at the time of writing but needs reworking.
-			// See: https://github.com/Kong/kubernetes-ingress-controller/issues/2130
 			p.logger.Errorf(err.Error())
 		}
 	}
